requirements/base.txt
Django==1.8.2
django-filter==0.10.0
djangorestframework==3.1.3
gnureadline==6.3.3
ipdb==0.8.1
ipython==3.1.0
Markdown==2.6.2
psycopg2==2.6.1
factory-boy==2.5.2
